Reorganize frontend module discussion into subsystems discussion
The Frontend modules are technically individual modules, but they do
too much work, glue too many other modules together, and are too high
in the module hierarchy to really be considered individual modules.

Instead, the discussion of the Frontend modules should be done with
the discussion of the Frontend subsystem as a whole.